Tools/zero_conf/pyproject.toml
[tool.poetry]
name = "zero_conf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mcrcon = "^0.7.0"
paho-mqtt = "^1.5.1"
getkey = "^0.6.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
